requirements.txt
asgiref==3.5.2
astroid==2.11.6
async-timeout==4.0.2
certifi==2022.6.15
cfgv==3.3.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
Deprecated==1.2.13
dill==0.3.5.1
distlib==0.3.4
Django==4.0.6
django-environ==0.9.0
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
filelock==3.7.1
identify==2.5.1
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
nodeenv==1.7.0
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.12.51
platformdirs==2.5.2
pre-commit==2.19.0
psycopg2==2.9.3
PyJWT==1.7.1
pylint==2.14.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.15.1
wrapt==1.14.1
